Fix memory setting issue when starting domain
By passing a flag to libvirt's API to apply the memory size change
to the domain's configuration file, the following issue was solved.

 -  fail to update the domain.memory on existing running vm 'Call
    to virDomainSetMemory failed' #1812
    https://github.com/vagrant-libvirt/vagrant-libvirt/issues/1812

The problem was caused by passing only memory size to #memory= method
in start_domain.rb. Specifying a flag in addition to the memory size
causes libvirt's API, virDomainSetMemoryFlags, to be called with that
flag, explicitly choosing whether to apply the size setting to the
running domain or the config file for that domain. The missing flag was
causing the above issue by attempting to change the memory size of
a running domain that did not exist.

Therefore, a list of possible flags is introduced, and a flag that
make changes to the domain's configuration file is selected and used.

Below are links to some related documents.

 -  https://ruby.libvirt.org/api/Libvirt/Domain.html#method-i-memory-3D

 -  https://libvirt.org/html/libvirt-libvirt-domain.html#virDomainSetMemoryFlags